separate import_utils module
